Remove PublishTrimmed property from plugin build script
It's currently not supported to use trimming for non-SCD apps. If the PublishTrimmed option is left in, you'll see an error like this:

```
error NETSDK1102: Optimizing assemblies for size is not supported for the selected publish configuration. Please ensure that you are publishing a self-contained app. [/home/vsts/work/1/s/src/plugins/ChronicleMunger/ChronicleMunger.csproj]
```

https://dev.azure.com/scorpdx/ironmunge/_build/results?buildId=368&view=logs&j=1cbb5fa8-227a-5271-32b9-169471ebf78b&t=ee99034a-8350-59b1-e746-a0270efb6ce5&l=45